fix: availableCapacity, suggestedCount 타입 Integer → long 변경

+++ hub-service/src/main/java/com/boxoffice/hubservice/stocktransfer/dto/response/SuggestedTransferResponseDto.java
@@ -8,7 +8,7 @@ public record SuggestedTransferResponseDto(
         UUID toHubId,
         String toHubName,
         BigDecimal distanceKm,
-        Integer availableCapacity,
-        Integer suggestedCount,
+        long availableCapacity,
+        long suggestedCount,
         List<AssignedCompanyResponseDto> companies
 ) { }
